troubleshooting

+++ backup/modules/logging.py
@@ -1,5 +1,6 @@
 import modules as mods
 from datetime import datetime
+from pathlib import Path
 
 def openLogFile(logpath,logfiletype,debug):
     """
@@ -22,7 +23,10 @@ def openLogFile(logpath,logfiletype,debug):
         logfilename = "updatelog.log"
 
     try:
-        logfile = open(logpath + logfilename,"w")
+        if Path(logpath + logfilename).exists():
+            logfile = open(logpath + logfilename,"a")
+        else:
+            logfile = open(logpath + logfilename,"w")
 
         if debug:
             print("[DEBUG] Logfile {} is aangemaakt".format(logfilename))
